Create chunks that exactly match the chunkSize.
Previously the size of the chunk was the smallest chunk equal to
or greater than the specified chunk size that could be made with
the smaller chunks of data provided from upstream. This commit
changes the behavious such that the chunk size is exactly as
specified, and any excess bytes are returned to the stream to be
included in the next chunk. As previously, the final chunk may be
smaller than the specified chunk size.